+++ inst/include/timedepGlobalStructs.h
@@ -31,7 +31,7 @@ typedef double (*RateFunctionsPtr) (double, void*);
 struct GlobalParameters
 {
   double tot_life;
-  int max_pop;
+  long unsigned int max_pop;
   int ancestors;
   int ancestor_clones;
   int num_sims;
@@ -66,7 +66,7 @@ struct clone
   // clone information
   std::string clone_id;
   int subclone_count;
-  int cell_count;
+  long unsigned int cell_count;
   int allele_count;
   int mut_count;
   int driver_count;
